fix: move api.js to utils folder
