+++ client/src/components/Message.js
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 import { FiPhoneCall } from "react-icons/fi";
 import { PiDogFill } from "react-icons/pi";
 import { toast } from "react-toastify";
-import { postData } from "../lib/api";
+import { postData } from "../utils/api";
 import { calculateAge } from "../utils/common";
 
 export default function Message({ refresh }) {
